requirements.txt
playwright==1.49.0
pytest-playwright==0.4.3
pytest==7.4.4
faker==37.8.0
pytest-html==4.1.1
pytest-xdist==3.8.0
pytest-rerunfailures==16.0.1
allure-pytest==2.15.0
